ExtractSignalV2.C: fix remaining BG Integral error computation
this fixes the cause of the displayed error message
'Error in <TF1Helper::IntegralError>: Last used fitter is not compatible with the current TF1'
when running ExtractSignalV2.C. The reason was that in the call of IntegralError
no covariance matrix for parameters of remaining BG functions were provided.
This lead to the error of the integral being set to 0.
With this fix errors on the raw yields will slightly increase.